Clean up convex MPC code
