+++ camel-canine/canine_controller/ControlMain/CMakeLists.txt
@@ -6,7 +6,12 @@ include_directories(
         ${EIGEN3_INCLUDE_DIRS}
         ${canine_util_SOURCE_DIR}/include
         ${PDcontroller_SOURCE_DIR}/include
+        ${convexMPC_SOURCE_DIR}/include
         ${GaitScheduler_SOURCE_DIR}/include
+        ${qpOASES_SOURCE_DIR}/include
+)
+link_directories(
+        ${qpOASES_SOURCE_DIR}/bin
 )
 
 add_library(${PROJECT_NAME} STATIC
@@ -14,3 +19,5 @@ add_library(${PROJECT_NAME} STATIC
 )
 
 add_library(canine_controller::${PROJECT_NAME} ALIAS ${PROJECT_NAME})
+target_link_libraries(${PROJECT_NAME}
+        canine_controller::convexMPC)
