+++ camel-canine/canine_controller/convexMPC/CMakeLists.txt
@@ -5,7 +5,6 @@ project(convexMPC LANGUAGES CXX)
 include_directories(
         include
         ${EIGEN3_INCLUDE_DIRS}
-        ${GaitScheduler_SOURCE_DIR}/include
         ${qpOASES_SOURCE_DIR}/include
         ${canine_util_SOURCE_DIR}/include
 )
@@ -14,15 +13,8 @@ link_directories(
 )
 
 add_library(${PROJECT_NAME} STATIC
-        src/Trajectory.cpp
-        src/qpsolver.cpp
-        src/MPCController.cpp
-)
+        src/MPCSolver.cpp
+        src/MPCController.cpp)
 
 add_library(canine_controller::${PROJECT_NAME} ALIAS ${PROJECT_NAME})
-target_link_libraries(${PROJECT_NAME}
-        PUBLIC
-        raisim::raisim
-        PRIVATE
-        canine_controller::GaitScheduler
-)
+target_link_libraries(${PROJECT_NAME} qpOASES)
