+++ camel-canine/canine_controller/convexMPC/include/convexMPC/MPCSolver.hpp
@@ -7,50 +7,50 @@
 
 #include <iostream>
 
-#include <raisim/World.hpp>
 #include <unsupported/Eigen/MatrixFunctions>
 #include <qpOASES.hpp>
 
 #include <canine_util/EigenTypes.hpp>
-#include <GaitScheduler/Gait.hpp>
+#include <canine_util/SharedMemory.hpp>
+
+#include "MPCdescription.hpp"
 
 using Eigen::Dynamic;
 
-class ConvexMPCSolver{
+class MPCSolver{
 public:
-    ~ConvexMPCSolver();
+    MPCSolver(const uint8_t& horizon);
+    ~MPCSolver();
 
-    void InitMatrix(raisim::Mat<3,3> bdyInertia);
-    void SetParameters(int Horizon, double Dt);
-    void SetWeights(Vec13<double> weight, double alpha);
-
-    void ResizeMatrix();
-    void GetMetrices(int *_mpcTable, raisim::VecDyn pos, raisim::VecDyn vel, raisim::Vec<3> footPosition[4]);
-    void qpSolver();
-    void SetTrajectory(double currentTime, GaitType currentGait);
-    void GetJacobian(Eigen::Matrix<double,3,3>& J, double hip, double thigh, double calf, int side);
+    void SetTrajectory();
+    void GetMetrices();
+    void SolveQP();
     void GetGRF(Vec3<double> f[4]);
-
-    Vec4<double> quat;
-    Vec3<double> p,v,w,q;
-
-    double stopPosX = 0.0;
+    void GetJacobian(Eigen::Matrix<double,3,3>& J, double hip, double thigh, double calf, int side);
 
 private:
-    void transformQuat2Euler(const Vec4<double>& quat, Vec3<double>& q);
-    void getStateSpaceMatrix(Eigen::Matrix<double,13,13>& Ac, Eigen::Matrix<double,13,12>& Bc, raisim::Vec<3> footPosition[4]);
-    void transformC2QP(Eigen::Matrix<double,13,13> A, Eigen::Matrix<double,13,12> B);
+    void initMatrix();
+    void resizeMatrix();
+    void updateRobotStates();
+    void getStateSpaceMatrix();
+    void transformC2QP();
+
     void transformMat2Real(qpOASES::real_t* dst, Eigen::Matrix<double,Dynamic,Dynamic> src, int16_t rows, int16_t cols);
 
 private:
-    uint8_t mHorizon;
-    double mDt;
-    double mAlpha;
-
-    Eigen::Matrix<double,3,3> I_world;
-    Eigen::Matrix<double,3,3> I_inv;
-
+    const double mDt;
+    const double mAlpha;
     Vec13<double> mWeightMat;
+    const uint8_t mHorizon;
+    Vec3<double> mBaseP;
+    Vec3<double> mBaseV;
+    Vec3<double> mBaseQ;
+    Vec3<double> mBaseW;
+
+    Eigen::Matrix<double,3,3> mBodyInertia;
+    Eigen::Matrix<double,3,3> mBodyInertiaInverse;
+
+    double stopPosX = 0.0;
 
     Eigen::Matrix<double,13,13> Ac;
     Eigen::Matrix<double,13,12> Bc;
