--- camel-canine/canine_controller/convexMPC/src/Trajectory.cpp
@@ -1,18 +1,0 @@
-//
-// Created by hs on 22. 7. 14.
-//
-
-#include <convexMPC/Trajectory.hpp>
-
-void LegTrajectory::updateTrajectory(double currentTime, double timeDuration){
-    mReferenceTime = currentTime;
-    mTimeDuration = timeDuration;
-}
-
-// 25 step period
-double LegTrajectory::getPositionTrajectory(double currentTime) {
-    double normalizedTime = (currentTime - mReferenceTime) / mTimeDuration;
-    int k = normalizedTime/0.125;
-    normalizedTime -= double(k) * 0.125;
-    return -25.6*pow(normalizedTime-0.0625,2)-0.27;
-}
