+++ camel-canine/canine_fsm/CMakeLists.txt
@@ -12,6 +12,11 @@ include_directories(
         ${canine_util_SOURCE_DIR}/include
         ${GaitScheduler_SOURCE_DIR}/include
         ${PDcontroller_SOURCE_DIR}/include
+        ${convexMPC_SOURCE_DIR}/include
+        ${qpOASES_SOURCE_DIR}/include
+)
+link_directories(
+        ${qpOASES_SOURCE_DIR}/bin
 )
 
 add_library(canine_fsm
