+++ camel-canine/canine_util/include/canine_util/RobotDescription.hpp
@@ -104,4 +104,11 @@ enum SIMULATION_STATE
     WITH_SIMULATION
 };
 
+enum GAIT_TYPE
+{
+    STAND = 0,
+    TROT
+};
+
+
 #endif //RAISIM_ROBOTDESCRIPTION_H
