+++ camel-canine/canine_util/src/Command.cpp
@@ -69,8 +69,8 @@ void Command::commandFunction()
             case CUSTOM_2:
             {
                 sharedMemory->controlState = STATE_TROT_REDAY;
-                sharedMemory->can1State = CAN_SET_TORQUE;
-                sharedMemory->can2State = CAN_SET_TORQUE;
+/*                sharedMemory->can1State = CAN_SET_TORQUE;
+                sharedMemory->can2State = CAN_SET_TORQUE;*/
                 break;
             }
             default:
