+++ camel-canine/canine_util/src/RobotMath.cpp
@@ -1,0 +1,28 @@
+//
+// Created by hs on 22. 10. 14.
+//
+
+#include <canine_util/RobotMath.hpp>
+
+template <class T>
+
+T t_min(T a, T b)
+{
+    if(a<b) return a;
+    return b;
+}
+
+template <class T>
+T sq(T a)
+{
+    return a*a;
+}
+
+void TransformQuat2Euler(const double* quat, double* euler)
+{
+    //edge case!
+    float as = t_min(-2.*(quat[1]*quat[3]-quat[0]*quat[2]),.99999);
+    euler[0] = atan2(2.f*(quat[2]*quat[3]+quat[0]*quat[1]),sq(quat[0]) - sq(quat[1]) - sq(quat[2]) + sq(quat[3]));
+    euler[1] = asin(as);
+    euler[2] = atan2(2.f*(quat[1]*quat[2]+quat[0]*quat[3]),sq(quat[0]) + sq(quat[1]) - sq(quat[2]) - sq(quat[3]));
+}
